Added sanitization, fixed bugs, refactored
:heavy_plus_sign: Added sanitize-html and validator to sanitize inputs before validating.
:bug: Reworked operations to use public ID instead of internal to make MongoDB and Cassandra work together smoothly. Internal ID will only be used for PostgreSQL operations where speed is key.
:bug: Fixed client not rendering on multiple actions, like leaving a server, deleting a server, joining channels on creation.
:art: Refactored code for components to be uniform.

+++ src/client/src/components/servers/server/Server.tsx
@@ -29,23 +29,32 @@ export default function Server({
 }: ServerProps) {
     const handleContextMenu = (e: MouseEvent<HTMLDivElement>) => {
         const isMember = server.users.some((u) => u.id === user.publicId)
-        const isOwner = server.ownerIds.length > 0
+        const isOwner = server.ownerIds.some((id) => id === user.publicId)
 
-        const options = isOwner
-            ? [
-                  {
-                      label: 'Delete server',
-                      action: () => deleteServer(server.id),
-                  },
-              ]
-            : isMember
-              ? [
-                    {
-                        label: 'Leave server',
-                        action: () => leaveServer(server.id),
-                    },
-                ]
-              : []
+        const copyServerId = () => {
+            navigator.clipboard.writeText(server.id)
+        }
+
+        const options: ContextMenuOption[] = [
+            {
+                label: 'Copy Server ID',
+                action: copyServerId,
+            },
+        ]
+
+        if (isOwner) {
+            options.push({
+                label: 'Delete Server',
+                action: () => deleteServer(server.id),
+                danger: true,
+            })
+        } else if (isMember) {
+            options.push({
+                label: 'Leave Server',
+                action: () => leaveServer(server.id),
+            })
+        }
+
         openContextMenu(e, options)
     }
